Rename and update item with no dividers
