+++ apps/cookbook/src/app/showcase/list-showcase/list-showcase.component.html
@@ -27,7 +27,10 @@
     <cookbook-code-viewer [html]="items"></cookbook-code-viewer>
   </div>
 
-  <cookbook-iphone src="/examples/list/with-items" showExternalLink="true"></cookbook-iphone>
+  <cookbook-iphone
+    src="/examples/list/with-items-no-dividers"
+    showExternalLink="true"
+  ></cookbook-iphone>
 </div>
 
 <div class="page-example">
